Added support for DEBUG in the entrypoint.sh

+++ src/docker/entrypoint.sh
@@ -10,6 +10,16 @@ fi
 if [ ! -z $GRID_INTERFACES ]; then
 	SYSTEM_PROPERTIES="$SYSTEM_PROPERTIES -DinMemoryGrid.interfaces=$GRID_INTERFACES"
 fi
-ln -f -s /run/secrets/churchclerk /home/config/application.properties
+
+if [ -f /run/secrets/churchclerk ]; then
+  ln -f -s /run/secrets/churchclerk /home/config/application.properties
+  if [ "$DEBUG" = "true" ]; then
+    cat /run/secrets/churchclerk
+  fi
+else
+  if [ "$DEBUG" = "true" ]; then
+    echo "No secret specified"
+  fi
+fi
 cd /home
 java $SYSTEM_PROPERTIES -cp /home -jar /home/app.jar $APP_ARG
